+++ Sources/KeyboardKit/_Deprecated/Feedback+Deprecated.swift
@@ -3,6 +3,9 @@ import Foundation
 @available(*, deprecated, renamed: "Feedback.Audio")
 public typealias AudioFeedback = Feedback.Audio
 
+@available(*, deprecated, renamed: "Feedback.Haptic")
+public typealias HapticFeedback = Feedback.Haptic
+
 public extension Feedback.Audio {
     
     @available(*, deprecated, renamed: "Feedback.AudioConfiguration")
@@ -15,4 +18,13 @@ public extension Feedback.Audio {
     static func custom(id: UInt32) -> AudioFeedback {
         customId(id)
     }
+}
+
+public extension Feedback.Haptic {
+    
+    @available(*, deprecated, renamed: "Feedback.HapticConfiguration")
+    typealias Configuration = Feedback.HapticConfiguration
+    
+    @available(*, deprecated, renamed: "Feedback.HapticEngine")
+    typealias Engine = Feedback.HapticEngine
 }