+++ Tests/KeyboardKitTests/Mocks/MockHapticFeedbackEngine.swift
@@ -9,18 +9,18 @@
 import MockingKit
 @testable import KeyboardKit
 
-class MockHapticFeedbackEngine: HapticFeedback.Engine, Mockable {
+class MockHapticFeedbackEngine: Feedback.HapticEngine, Mockable {
 
     var mock = Mock()
     
     lazy var prepareRef = MockReference(prepare)
     lazy var triggerRef = MockReference(trigger)
     
-    override func prepare(_ feedback: HapticFeedback) {
+    override func prepare(_ feedback: Feedback.Haptic) {
         call(prepareRef, args: (feedback))
     }
 
-    override func trigger(_ feedback: HapticFeedback) {
+    override func trigger(_ feedback: Feedback.Haptic) {
         call(triggerRef, args: (feedback))
     }
 }